Fixed the PCAP python simulation tests.

+++ tests/test_python_sim_timing.py
@@ -10,6 +10,7 @@ else:
 import sys
 import os
 import imp
+import numpy
 
 import unittest
 
@@ -117,7 +118,7 @@ def load_tests(loader=None, standard_tests=None, pattern=None):
 
                 # Check all outputs have correct field values
                 for name in outputs:
-                    expected = int(outputs[name], 0)
+                    expected = numpy.int32(int(outputs[name], 0))
                     actual = getattr(block, name)
                     assert actual == expected, \
                         "%d: Attr %s = %d != %d" % (
